Changed: Todo key is now Long.

+++ src/main/kotlin/com/example/todoApp/Todo.kt
@@ -1,3 +1,3 @@
 package com.example.todoApp
 
-data class Todo(val id: Int, val text: String)
+data class Todo(val id: Long, val text: String)
